+++ domain/src/main/java/com/dawoud/domain/model/GeneralMovieModel.kt
@@ -5,14 +5,14 @@ data class GeneralMovieModel(
     val results : List<MovieModel>
 )
 data class MovieModel(
-    val id : Int,
-    val adult : Boolean,
-    val backdrop_path : String?,
-    val original_language :String,
-    val original_title :String,
-    val overview :String,
-    val popularity :Float,
-    val poster_path :String?,
-    val release_date :String,
-    val title :String
+    val id: Int,
+    val adult: Boolean,
+    val backdrop_path: String?,
+    val original_language:String,
+    val original_title:String,
+    val overview:String,
+    val popularity: Double,
+    val poster_path:String?,
+    val release_date:String,
+    val title:String
 )
